Move bucket logic to utils file (#1592)
* rm bjs stuff for now

* lint
