+++ frontend/src/cards/TableCard.tsx
@@ -15,58 +15,21 @@ import {
   MetricId,
   VariableConfig,
   getPer100kAndPctShareMetrics,
-  VariableId,
 } from "../data/config/MetricConfig";
 import { exclude } from "../data/query/BreakdownFilter";
-import {
-  NON_HISPANIC,
-  RACE,
-  UNKNOWN,
-  UNKNOWN_RACE,
-  UNKNOWN_ETHNICITY,
-  ALL,
-  BROAD_AGE_BUCKETS,
-  DECADE_PLUS_5_AGE_BUCKETS,
-  VOTER_AGE_BUCKETS,
-  AGE_BUCKETS,
-  ASIAN_NH,
-  NHPI_NH,
-  API_NH,
-} from "../data/utils/Constants";
-import { Row } from "../data/utils/DatasetTypes";
+import { RACE } from "../data/utils/Constants";
 import MissingDataAlert from "./ui/MissingDataAlert";
 import Alert from "@material-ui/lab/Alert";
 import Divider from "@material-ui/core/Divider";
-import {
-  UHC_API_NH_DETERMINANTS,
-  UHC_DECADE_PLUS_5_AGE_DETERMINANTS,
-  UHC_DETERMINANTS,
-  UHC_VOTER_AGE_DETERMINANTS,
-} from "../data/variables/BrfssProvider";
 import { urlMap } from "../utils/externalUrls";
-import { shouldShowAltPopCompare } from "../data/utils/datasetutils";
+import {
+  getExclusionList,
+  shouldShowAltPopCompare,
+} from "../data/utils/datasetutils";
 import styles from "./Card.module.scss";
-
-const showAllGroupIds: VariableId[] = [
-  "women_state_legislatures",
-  "women_us_congress",
-];
-
-// only for UHC variables
-const ALL_UHC_DETERMINANTS = [
-  ...UHC_DECADE_PLUS_5_AGE_DETERMINANTS,
-  ...UHC_VOTER_AGE_DETERMINANTS,
-  ...UHC_DETERMINANTS,
-];
 
 /* minimize layout shift */
 const PRELOAD_HEIGHT = 698;
-
-export interface TableCardProps {
-  fips: Fips;
-  breakdownVar: BreakdownVar;
-  variableConfig: VariableConfig;
-}
 
 // We need to get this property, but we want to show it as
 // part of the "population_pct" column, and not as its own column
@@ -75,52 +38,23 @@ export const NEVER_SHOW_PROPERTIES = [
     ?.secondaryPopulationComparisonMetric,
 ];
 
+export interface TableCardProps {
+  fips: Fips;
+  breakdownVar: BreakdownVar;
+  variableConfig: VariableConfig;
+}
+
 export function TableCard(props: TableCardProps) {
   const metrics = getPer100kAndPctShareMetrics(props.variableConfig);
-  const current100k = props.variableConfig.metrics.per100k.metricId;
-
-  // choose demographic groups to exclude from the table
-  let exclusionList = [];
-
-  if (!showAllGroupIds.includes(props.variableConfig.variableId)) {
-    exclusionList.push(ALL);
-  }
-
-  // across all variables
-  if (props.breakdownVar === "race_and_ethnicity") {
-    exclusionList.push(NON_HISPANIC);
-  }
-
-  if (
-    ALL_UHC_DETERMINANTS.includes(current100k) &&
-    props.breakdownVar === "race_and_ethnicity"
-  ) {
-    UHC_API_NH_DETERMINANTS.includes(current100k)
-      ? exclusionList.push(ASIAN_NH, NHPI_NH)
-      : exclusionList.push(API_NH);
-  } else if (
-    ALL_UHC_DETERMINANTS.includes(current100k) &&
-    props.breakdownVar === "age"
-  ) {
-    // get correct age buckets for this determinant
-    let determinantBuckets: any[] = [];
-    if (UHC_DECADE_PLUS_5_AGE_DETERMINANTS.includes(current100k))
-      determinantBuckets.push(...DECADE_PLUS_5_AGE_BUCKETS);
-    else if (UHC_VOTER_AGE_DETERMINANTS.includes(current100k))
-      determinantBuckets.push(...VOTER_AGE_BUCKETS);
-    else if (UHC_DETERMINANTS.includes(current100k))
-      determinantBuckets.push(...BROAD_AGE_BUCKETS);
-
-    // remove all of the other age groups
-    const irrelevantAgeBuckets = AGE_BUCKETS.filter(
-      (bucket) => !determinantBuckets.includes(bucket)
-    );
-    exclusionList.push(...irrelevantAgeBuckets);
-  }
 
   const breakdowns = Breakdowns.forFips(props.fips).addBreakdown(
     props.breakdownVar,
-    exclude(...exclusionList)
+    exclude(
+      ...(getExclusionList(
+        props.variableConfig,
+        props.breakdownVar
+      ) as string[])
+    )
   );
 
   let metricConfigs: Record<string, MetricConfig> = {};
@@ -161,16 +95,11 @@ export function TableCard(props: TableCardProps) {
       }
     >
       {([queryResponse]) => {
-        let dataWithoutUnknowns = queryResponse.data.filter(
-          (row: Row) =>
-            row[props.breakdownVar] !== UNKNOWN &&
-            row[props.breakdownVar] !== UNKNOWN_RACE &&
-            row[props.breakdownVar] !== UNKNOWN_ETHNICITY
-        );
+        let data = queryResponse.data;
 
         if (shouldShowAltPopCompare(props)) {
           // This should only happen in the vaccine kff state case
-          dataWithoutUnknowns = dataWithoutUnknowns.map((item) => {
+          data = data.map((item) => {
             const {
               vaccine_population_pct,
               acs_vaccine_population_pct,
@@ -228,7 +157,7 @@ export function TableCard(props: TableCardProps) {
             {!queryResponse.dataIsMissing() && (
               <div className={styles.TableChart}>
                 <TableChart
-                  data={dataWithoutUnknowns}
+                  data={data}
                   breakdownVar={props.breakdownVar}
                   metrics={Object.values(metricConfigs).filter(
                     (colName) => !NEVER_SHOW_PROPERTIES.includes(colName)
